feat: Implement minimal plugin architecture with water tracking MVP
Major architectural refactor to support plugin-based data collection:

Core Changes:
- Simplified DataPoint model for MVP, removing location fields
- Created minimal Plugin interface focused on manual data entry
- Implemented event-driven architecture with EventBus for loose coupling
- Fixed DataPointEntity location and database schema

Plugin System:
- Added WaterPlugin with quick-add functionality (Glass/Bottle/Liter/Custom)
- Added CounterPlugin for testing the plugin system
- Created PluginManager for plugin lifecycle and state management
- Implemented QuickAddConfig for declarative UI generation

Data Layer:
- Updated DataRepository with proper entity conversions
- Fixed database entity field mappings (valueJson, metadataJson)
- Added flexible JSON storage for future expansion
- Included reserved columns in DataPointEntity for future features

UI Updates:
- Created QuickAddBottomSheet for plugin data entry
- Updated DashboardScreen with plugin tiles and quick-add support
- Fixed DataScreen to work without location data
- Added plugin-based filtering in data view

Bug Fixes:
- Resolved duplicate class definitions in wrong file locations
- Fixed PluginManager references to non-existent methods
- Updated all UI components to use new signatures
- Corrected database schema version conflicts

This MVP provides a solid foundation for future enhancements including
voice input, AI analysis, P2P sync, and blockchain integration while
keeping the current implementation simple and focused.
